Network.TextViaSockets

Description

Simple line-based text communication via sockets.

Documentation

data Connection

A connection for sending and receiving Text lines.

Connect to a server

connectTo :: HostName -> ServiceName -> IO Connection

Connect to the given host and service name (usually a port number).

If the connection cannot be established, this action will be retried using an exponential back-off strategy, until the maximum number of tries is reached.

Start a server

acceptOn :: PortNumber -> IO Connection

Accept byte-streams by serving on the given port number. This function will block until a client connects to the server.

If the connection cannot be established, this action will be retried using an exponential back-off strategy, until the maximum number of tries is reached.

acceptOnSocket :: Socket -> IO Connection

Like acceptOn but it takes a bound socket as parameter.

getFreeSocket :: IO Socket

Get a free socket from the operating system.

Sending and receiving data

getLineFrom :: Connection -> IO Text

Read a text line from the given connection.

This function might throw an BlockedIndefinitelyOnSTM exception if the connection to the server is closed, so users of this function should check for this.

receiveMsgs :: Connection -> Int -> IO [Text]

Receive n messages.

putLineTo :: Connection -> Text -> IO ()

Put a text line onto the given connection.

sendMsgs :: Connection -> [Text] -> IO ()

Send a list of messages.

Closing the connection

close :: Connection -> IO ()

Close the connection.